fix: update parameter value (#4955)

+++ perception/image_projection_based_fusion/include/image_projection_based_fusion/roi_cluster_fusion/node.hpp
@@ -43,13 +43,13 @@ protected:
   bool use_iou_{false};
   bool use_cluster_semantic_type_{false};
   bool only_allow_inside_cluster_{false};
-  float roi_scale_factor_{1.1f};
-  float iou_threshold_{0.0f};
-  float unknown_iou_threshold_{0.0f};
+  double roi_scale_factor_{1.1};
+  double iou_threshold_{0.0};
+  double unknown_iou_threshold_{0.0};
   const float min_roi_existence_prob_ =
     0.1;  // keep small value to lessen affect on merger object stage
   bool remove_unknown_;
-  float trust_distance_;
+  double trust_distance_;
 
   bool filter_by_distance(const DetectedObjectWithFeature & obj);
   bool out_of_scope(const DetectedObjectWithFeature & obj);
